* gprbuild_ug.texi: Update section Compilation of one or several sources
* buildgpr.adb:
Remove special case of gprbuild -c with a main specified on the command
line.

Implements H707-017

[gprbuild] No special case of gprbuild -c with main on command line

The special case of gprbuild -c with main(s) specified on the command
line is removed. The behavior is now the same, whether the mains are
specified on the command line or in attribute Main.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@126503 936e1b1b-40f2-da11-902a-00137254ae57