* buildgpr.adb: (Post_Compilation_Phase): If If -dn was used, indicate to gprbind that  the temporary response file, if created, should not deleted.  Send the Toolchain Versions of each language when they are declared.
* gprbind.adb:
Take into account new sections TOOLCHAIN VERSION and
DELETE TEMP FILES.
Create a response file only when the GNAT version is 6 or later.
Do not delete temporary response file if DELETE TEMP FILES is False.

* gprexch.ads:
(Binding_Section): Add two new labels: Toolchain_Version and
 Delete_Temp_Files.

For H709-016

[gprbuild] No response file for gnatbind if GNAT version not recent enough

If there are too many ALI files to send to gnatbind, gprbind creates a
response file. However, this does not work if the version of GNAT is not
recent enough. This patch ensures that a response file for gnatbind is
only created if gnatbind will accept it.
Also, if gprbuild is invoked with -dn, do not delete the temporary
gnatbind response file, if one has been created.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@126686 936e1b1b-40f2-da11-902a-00137254ae57