* linker.xml: Declare a Library_Partial_Linker for HP-UX and Windows For H116-039 and H712-006
* buildgpr.adb:
(Build_Library): Always put in exchange file the library partial linker
 information, if attribute Library_Partial_Linker is declared, not only
 for static libraries.
For H116-039 and H712-006

* gprbind.adb:
Invoke gnatbind with -F to always check elaboration flags, to avoid
repeated elaboration with Stand-Alone Libraries.
Found while investigating H712-006

* gprlib-build_shared_lib-nosymbols.adb:
(Build): If there is a partial linker, build the shared library in chunks
 if needed.
For H712-006

* gprlib.adb:
(Partial_Name): New function to get the file name of partial link object
 files.
(Gprlib): For static SALs, if partial linker is used, build in chunks
 if needed.
For H116-039

When attribute Library_Partial_Linker is declared, when the partial
linker is used, link the library in chunks, if the command line is
too long to build in one chunk.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@126893 936e1b1b-40f2-da11-902a-00137254ae57