* gprbuild_ug.texi: Add documentation for new gprbind switch --gnatbind_path=
* gprbind.adb:
Add new switch --gnatbind_path=
When this switch is not used and gnatbind cannot be found at the same
location as the compiler, look for gnatbind on the path.

For I330-024

[gprbuild] Improvements for gnatbind
A new switch --gnatbind_path= is introduced, so that the proper gnatbind
can be specified by the user, when it is not the one at the same location
as the Ada compiler. In addition, when this switch is not specified and
gnatbind is not found at the same location as the Ada compiler, gprbind
now looks for gnatbind on the path.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@142173 936e1b1b-40f2-da11-902a-00137254ae57