* linker.xml: For Linux and Solaris, when C++ is a language, do not bind with -shared by default, but link with -shared-libgcc in the Linker'Required_Switches.
* buildgpr.adb:
(Add_Rpath): Do nothing if Path is empty
(Linking_Phase): Make sure that when a directory is added to the run path
 option or when -shared-libgcc is used, the shared libgcc directory is in
 the run path option.

* gpr_util.adb:
(Shared_Libgcc_Dir): New function to get the shared libgcc directory from
 the Ada runtime directory.

* gpr_util.ads:
(Shared_Libgcc_Dir): New function to get the shared libgcc directory from
 the Ada runtime directory.

* gprlib.adb:
(Add_Rpath (String)): New procedure
(Add_Rpath (String_Access)): Do not add a directory if it is already in
 Rpath.
(Gprlib): When adding the Ada runtime directory to the run path option,
 also add the shared libgcc directory.

Fixes IA18-001

When a shared library is linked, put the proper shared libgcc directory
in the run path option. When an executable is linked with -shared-libgcc,
make sure that the shared libgcc directory is in the run path option.
On Solaris and Linux, do not use -shared to bind by default, but always
link with -shared-libgcc.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@151134 936e1b1b-40f2-da11-902a-00137254ae57