* buildgpr.adb: (Bind_Language): Instead of looking for the toolchan versions of the  languages of the project, use only the one of the current language,  if there is one, as this may not be a language of the project.
* gprbind.adb:
(GNAT_Version_Set): New Boolean variable, default to False, set to True
 when the GNAT version is found in the input exchange file.
Assume the GNAT version is ancient when GNAT_Version_Set is False.

* gprlib.adb:
(GNAT_Version_Set): New Boolean variable, default to False, set to True
 when the GNAT version is found in the input exchange file.
Assume the GNAT version is ancient when GNAT_Version_Set is False.
Do not use a response file when the GNAT version is not set or is 3.xx or
5.xx.

Continuation of K421-022




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@172722 936e1b1b-40f2-da11-902a-00137254ae57