* gpr_util.adb: (Process_ALI_Closure_Deps): New function to be used when dependency kind  is ALI_Closure.
* gprbuild-compile.adb:
(Wait_For_Available_Slot): Use Phase_2_ALI when Dependency_Kind is
 ALI_Closure.

Part of K513-011
Implementation of dependency kind ALI_Closure, to be used by gnatprove.
When dependency kind is ALI_Closure, a source is recompiled if there is
a time stamp discrepancy in any of the sources of its closure, not just
those it depends directly on.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@186994 936e1b1b-40f2-da11-902a-00137254ae57