Add gprinstall support for copying DLL into bin directory on Windows.
This is needed for the DLLs to be placed into the bin directory
for them to be found in the PATH.

The option --no-lib-in-exec has been introduced to disbale this
default behavior. This is needed to control which version of the
DLLs is to be placed into the PATH when multiple builds are
installed.

Update tests accordingly, add one specific output for Windows.

Continued work for H729-014.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@192144 936e1b1b-40f2-da11-902a-00137254ae57