ravencar-cert-rtp: do not link against rtpCert.o
This part is not deactivated by -nostdlib, and as Workbench links
against the same object we run into conflicts. In order to keep
gprbuild working when not invoked from Workbench, remove -nostdlib and
-nostartfiles as well; let it link (implicitly) against the system
libc.a. This implies that we have two different link strategies
depending on the place from which gprbuild is invoked (command line or
Workbench); whether or not this is OK for certification is still to be
discussed.

[L921-009]

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@195522 936e1b1b-40f2-da11-902a-00137254ae57