Support distributed cross-compilation and configuration.
This patchset adds the ability to compile remotely with different
installation path for the compiler tool chain. The information
given to the slave is the language and the target of the file
to compile. From these information the slave will choose the
appropriate compiler on the path.

Another consequence of this enhanced implementation is that it is
possible to use the distributed compilation support with cross-compiler.
That is, it is possible to specify a target to gprbuild (option --taget)
and still use the distributed compilation.

If the toolchain supports it we can even compile locally on Windows and
use slaves running on GNU/Linux with Windows cross-compilers.

Continued work for K826-028.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@204342 936e1b1b-40f2-da11-902a-00137254ae57