Fix race condition in Wait_Local.
We make sure that a local slot is freed before registering the result.
Not doing so a new compilation could be started before freeing a local
resource, so a remote compilation job would be sent. But if we have reached
the maximum number of remote compilation we will fail to find a free
slot on a slave.

Fixes M326-031.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@207332 936e1b1b-40f2-da11-902a-00137254ae57