Add support for a builder set.
This is another step toward support for multiple user for the distributed
compilation. This patch introduces the a set of builder and a selector
to check for data availability on the sockets. So in a sense it already
support multiple user but the compilation jobs are not limited to the
number of cores per instance.

Part of M409-036.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@209209 936e1b1b-40f2-da11-902a-00137254ae57