Rename Full_Path_Tag to WD_Path_Tag.
This is the tag to identify the root path of the working directory.
The renaming is done as another tag will be used for the compiler
directory.

The tag is now set to '<1>' instead of just '<>'. At the same time
Get_Args has been fixed to not suppose a tag is 2 characters long.

Continued work for K826-028.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@209217 936e1b1b-40f2-da11-902a-00137254ae57