Rewrite also the compiler path.
The path for the compiler on the build master and on the slaves may not
be the same. We need to handle this properly by rewriting in the mapping
file the path to sources files outside of the working directory. This is
the case for all projects used that are not part of the current compilation
but are installed with the compiler.

Part of K826-028, found while working on M409-036.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@209221 936e1b1b-40f2-da11-902a-00137254ae57