pikeos: change entry point for the native personality
In the SYSGO toolchain, the name of the entry point is different
between APEX and native personalities: _p4_entry on native, _begin on
APEX. It seems that we can use _begin for both though. _begin is
actually defined in SYSGO's libapex, so we cannot change the name.
But _p4_entry seems to be just a name convention.

Changing the entry point allows the APEX link options to be a superset of
the native ones, and to have only one rts for both personalities.

Part of M611-012

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@211232 936e1b1b-40f2-da11-902a-00137254ae57