(Scanner): New scanner instantiated with empty error procedures  to avoid any error report when computing a checksum. (Process_ALI_Deps): Use new Scanner (Process_ALI_Closure_Deps): Use new Scanner
Fixes M625-043

When scanning a source to get its checsk sum and that source needs to be
preprocessed, gprbuild was reporting errors because of the preprocessing
directives. These arrors are no longer reported and gprbuild now behaves
correctly when invoked with switch -m, even with sources that need to be
preprocessed.



git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@211958 936e1b1b-40f2-da11-902a-00137254ae57