Add new routine to compute a slave environment based on project status.
This new routine returns an MD5 sum of the sorted command line parameters
and project variables. The values are sorted as we want the same MD5
value (slave environment) to be generated even if the argument are not
passed on the same order on the command line.

Part of M806-019.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@215388 936e1b1b-40f2-da11-902a-00137254ae57