Skip some options when computing the slave environment.
This will ensure that the same environment will be created for gprbuild
and gprclean as we do not take into account some options that are not
passed to the compiler. Those options are just changing the gprbuild
or gprclean output (-d for progress) or the behavior (-p for creating
directory).

Continued work for M806-019.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@215434 936e1b1b-40f2-da11-902a-00137254ae57