Use a new option to create destination directory.
The gprinstall tool was using the Setup_Projects option to control the
creation of destination directories. The problem is that this option is
already used to create missing directories in the build location.

At this point gprinstall use a separation option to control the creation
of directories at the installation point.

For MA29-049.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@218303 936e1b1b-40f2-da11-902a-00137254ae57