Adds all system libraries into the generated Linker_Options.
A system library is a library with no source and set as externally built.
If a project has such library in its dependencies it is to bring some
symbols needed at link time. gprinstall now properly handle this case and
adds the system libraries into the generated Linker_Options attribute of
the linker package.

Par of MA30-041.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@218487 936e1b1b-40f2-da11-902a-00137254ae57