Never copy the .d from the slaves when finalizing slaves.
The .d file has already been copied during compilation and at this
time the paths for the dependencies have been rewritten to conform
to build master directory structure. If we copy it again it will
ends up with wrong paths and will triger recompilation of the
sources built on the slaves.

At the same time we do not copy the .ali (already copied during
the build process). And we copy the .gli files.

For MC05-008.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@220493 936e1b1b-40f2-da11-902a-00137254ae57