Fix install of aggregated libraries.
This is a better fix for N421-002 latest commit.

First we do not iterate over the locally removed files. This avoid
one test and simplify the code.

Second we properly check that installed sources for an aggregated library
are part of one of the aggregated projects.

Continued work for N421-002.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@226909 936e1b1b-40f2-da11-902a-00137254ae57