* src/gprbuild.ads: Add necessary with's, and make some of them private.
* src/gprbuild-compile.adb:
Use appropriate package to find compiler switches. It was package Compiler. Now
it's package Compiler by default, but can be specified by --compiler-pkg-subst.

* src/gprbuild-main.adb:
Parse new switches, and make them incompatible with certain other switches.
Do a pre-pass to detect the new switches; otherwise, the compatibility check
would fail if they are given late in the command line.
Add new switches to the --help message.

* src/gpr_util.ads,src/gpr_util.adb:
Keep a per-language table for --compiler-subst, and a variable for
--compiler-pkg-subst.

* doc/gprbuild_ug.texi:
Document the new switches.

Part of N714-001.

This patch implements two new switches: --compiler-subst and
--compiler-pkg-subst, which are needed for proper implementation of
--incremental in ASIS tools in the presence of project files.


git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@233847 936e1b1b-40f2-da11-902a-00137254ae57