Wait for an ACK after an end-of-compilation message.
This is to ensure that a second gprbuild run just after will not
report an error for re-using the same build environment. Indeed there
was a small window between the time the send-of-compilation was received
and the builder was actually removed form the gprslave list. During this
period a build master using the same build environment was improperly
rejected.

For NA31-028.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@233920 936e1b1b-40f2-da11-902a-00137254ae57