Add support for out-of-tree build.
A new option --in-place has been introduced to be able to build
a project file out of its source tree. That is all objects, libraries,
and ali files are relocated to the directory where gprbuild (or
related tools) is invoked when using the --in-place option.

Three tools have been added this support, gprbuild, gprinstall and
gprclean to build, install or clean build artefacts out of the source
tree.

Add corresponding regression tests.

For O325-017.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@237655 936e1b1b-40f2-da11-902a-00137254ae57