Add support for out-of-tree build.
This is an updated version with two options to make the implementation
clearer. The option --relocate-build-tree replace the --in-place one
and an optional directory can be specified for the build directory. So
this implementation removes the limitation of beeing able to build
only in the current working directory. The new option --root-dir is
to be used to specified the artifacts root-dir for proper relocation.

Three tools have been added this support, gprbuild, gprinstall and
gprclean to build, install or clean build artefacts out of the source
tree.

Update corresponding regression tests.

Continued work for O325-017.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@237725 936e1b1b-40f2-da11-902a-00137254ae57