Use install-name for grouping include/lib directories.
If option install name is specified we now use it a suffix for the include
and lib directories. This will ensure grouping of all projects under a
common directory when a recursive install is made for example.

Install name was already used to create a single manifest file in such
case. So making it possible to uninstall with a single gprinsall command.
This grouping has been extended to include and lib directories to avoid
polluting those directories with lot of entries.

Updated expected output accordingly.

Motivated by L626-014.
For O603-011.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@239292 936e1b1b-40f2-da11-902a-00137254ae57