Do nothing if the builder is not found for a given request.
When a request from a build master is received if no builder are found
to handle this request in the slave just skip the request.

This can happen if the builder has been terminated in the mean-time. Note
that this just avoids some error message but in no case a critical error
of the slave.

Motivated by O720-008.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@240556 936e1b1b-40f2-da11-902a-00137254ae57