When a builder is interrupted, kill all corresponding processes.
This ensure that no job will remain running when a builder is terminated.
This is also exactly what is done on GPRbuild non distributed when a
compilation is interrupted with ctrl-c.

It also fix a small race condition if the same compilation (same file) is
sent just after an interrupted build.

Part of O720-008.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@240899 936e1b1b-40f2-da11-902a-00137254ae57