Introduce a new SI message kind in distributed protocol.
This new kind is used when a compilation process is killed (SI for
signal). This is needed as when gprbuild is stopped (ctrl-c) we do
not want to send back any acknowledgement after closing the builder.

This avoids some lock/exception on the slave. At the same time
this patch prepares work for killing properly processes.

Part of O903-042.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@241842 936e1b1b-40f2-da11-902a-00137254ae57