Fix the circuitry for kill processes belonging to a specific builder.
A killed status is now recorded for each job. When a killed job is found
we return No_Job. This ensure that gprslave do not try to send over any
artifacts.

We detect also the case where a process is started and at the time it
is registered the builder has been killed. In this case we record this
job as a killed one and we do kill the corresponding process. We do not
needed the compilation to continue and do not want it to continue actually
to be sure that it won't overwrite an artifact generated by a new builder
restarted just after.

For O903-042.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@241843 936e1b1b-40f2-da11-902a-00137254ae57