GPRinstall install all sources by default now.
This was the case with -a option before. This option is kept for
compatibility. The new switch -m has been introduced to revert to the
previous behavior. That is, -m will copy the minimal set of sources
for the installation to be used.

Update the documentation and the regression test accordingly.

For O924-030.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@242019 936e1b1b-40f2-da11-902a-00137254ae57