* gprbind.adb: Take into account Nothing_To_Bind
* gprbuild-link.adb:
(Link_Main): Do not relink if the binder exchange file indicates that
 there is nothing to bind.

* gprbuild-post_compile.adb:
(Post_Compilation_Phase.Bind_Language): When there is nothing to bind
 indicate it in the binder generated file.

* gprexch.ads:
(Bindng_Section): Add Nothing_To_Bind

For OA25-001

When there is Ada code, but there is nothing to bind (for example when
the main is in C and it has no Roots), gprbuild was always relinking,
even though it was not necessary. With this patch, gprbuild will no longer
relink when it is not necessary.




git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@243138 936e1b1b-40f2-da11-902a-00137254ae57