Refine implementation to support for restricting export symbols out of SALs.
This new implementation introduces an new table and linker section which
contains the objects to be used as interfaces to restrict symbols. This is
cleaner and makes less assumption about where the foreign language objects
are. At the same time it let the gprlib tool simple without any knowledge
about sources and object code.

Add corresponding regression test.

For O331-015.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@243178 936e1b1b-40f2-da11-902a-00137254ae57