Fix small race condition when a new compilation is started.
That is there was a small race condition between the spawn of the
compilation and the job registering. Between the process could
be terminated and retreived by the Wait_Completion task. But
at this point the job is not yet registered and can't be found, and
so there is no way to know the actual builder to talk to.

Motivated by OB20-022.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@244128 936e1b1b-40f2-da11-902a-00137254ae57