Add support for hash string in GPRbuild and GPRslave.
This can be used to ensure the same environment is used on both
side (slave and master).

   $ gprslave --hash=$(echo $ADA_PROJECT_PATH | shasum)

   $ gprbuild --hash=$(echo $ADA_PROJECT_PATH | shasum) --distributed=...

The string has no hard coded meaning. User's can fill it with whatever
is required to ensure that both side are using the same projects,
libraries versions, etc.

Add corresponding regression test.

For P112-012.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@245440 936e1b1b-40f2-da11-902a-00137254ae57