Fix support for Included_Artifact_Patterns.
The support for this attribute has been lost during the move from the
rsync to the adhoc protocol. This patch adds back the support for
this attribute which make it possible to copy back from the slaves some
specific artifacts generated during the compilation.

Add corresponding regression test.

For P112-067.

git-svn-id: svn+ssh://svn.eu.adacore.com/Dev/trunk/gprbuild@245518 936e1b1b-40f2-da11-902a-00137254ae57