Do not free compilation options (P503-040)
On Windows, when a compilation option is modified so that backslashes
are doubled, the original option was freed. However, the original
option is recorded in All_Options and it will be dereferenced later.
So, it should not be freed.
This should explain random failures of P215-047 on some Windows
platforms.

Change-Id: Idc15ea6ae94f1a07fce45e32f36081ab77c4586c