Add support for GPRinstall --side-debug option.
This option can be used to write the debug symbols in a side file instead
of keeping them in the executable or shared library. This is achieved
using objcopy and strip executable.

Note that this also supports cross compilation/installation if the
corresponding objcopy and strip are installed.

For P425-023.

Change-Id: I0d2283c7793a08e39a1adadab5639bcf1f25dac8