Rework the Communication_Channel and Command objects.
The goal is to fix a memory leak but this patch is using a controlled
object to achieve this goal. The previous implementation was not
releasing all memory and worse it could trigger a memory issue by
freeing the same memory chunk multiple times.

Finally, this implementation is cleaner as it avoids customized code
for the release procedure.

Note that a code refactoring has been made to share the process id
between the Slave and Process children.

Continued work for PB24-031, PB24-030 and PB25-017.

Change-Id: Ifaebfbac5c0b9cff07445254b41a26fdd13a390e