linker.xml: limit use of Wind_Base to platforms where it is defined
This fixes an issue which appeared on arm-vx7 following a change
to this file which added "GCC_Pro" as one of the possible compilers
for one of the entries which add a Linker'Required_Switches
directive relative to the value of Wind_Base. This entry is
currently limited to all VxWorks targets, which is not enough,
since Wind_Base is only defined for all VxWorks targets except
VxWorks 7.

This patch brings the target-based filter in line with the filter
used to determine whether to define Wind_Base or not.

Part of PA27-037.

Change-Id: I1d899a670b169cf5e3107a00ea8d662f7860e3ec