cross.xml: specify RTP base with GCC_Pro compiler also
This is a consequence of a recent gprconfig enhancemnt which
allows us to use AdaCore's C compiler on VxWorks when placed
ahead of the WRS one.  Now that we can use AdaCore's compiler,
it exposes some holes in our knowledge base regarding that compiler

In this case, we're missing the linker arguments which allows
us to specify the RTP base address.

Part of PC08-065.

Change-Id: Ie67dad973ea8e83d7f829f55cb1deec6873c5e3e