Reorganize sources to share some code with LibGPR2 and GPRremote.
This is motivated by the fact that many routines will be needed for the
new LibGPR2 project. This is a temporary solution but for a long period
of time as the migration to this new technology will take quite some
time.

All packages moved from the main src directory to LibGPR have been
placed under the GPR parent package for clarity. Also the different
packages containing utility routines have been merged together.

Part of OB10-014.

Change-Id: I2c1d368267721a7c5247f27647e54c3a46b07490