Rewrite bootstrap.sh
This rewrite was primarily motivated by the need to fix an issue where
GNATMAKEFLAGS was causing failure when called on the command-line:

    ./bootstrap.sh: unknown option GNATMAKEFLAGS=-bargs -E; try ./bootstrap.sh --help
    ./bootstrap.sh: unknown option --GNATMAKEFLAGS=-bargs -E; try ./bootstrap.sh --help

However, I took the opportunity to clean up the script, removing several
flaws while producing something more idiomatic while still retaining
POSIX compliance.

A new feature I've included is support for staged installs [1] via the
DESTDIR environment to make my life a little easier and allowing for you
to eventually honour GNU's install directories [2] in the rest of the code
base without needing to change the semantics of this script.

I've also removed the surprising behaviour of passing environment
variables in as command-line arguments to scripts by making them
explicitly environmental.

Lastly I've removed the assumptions around $0 and where to find the
directory the script was executed from.  This is completely unreliable
and should never be used. [3]

I think the result is cleaner, clearer and more easily adapted to new
requirements

1. https://www.gnu.org/prep/standards/html_node/DESTDIR.html
2. https://www.gnu.org/prep/standards/html_node/Directory-Variables.html
3. http://mywiki.wooledge.org/BashFAQ/028

Signed-off-by: Earnestly <zibeon@googlemail.com>
Change-Id: If7b133f1a5b03715f300c66ad578a3cf0d3ce717
(no-rh-check)