New switch --no-complete-output (Q123-019)
New gprbuild --no-complete-output (or equivalent -n) is added.
When used, gprbuild deletes the .stderr and .stdout files, if they
exist, and does not redirect the standadr output and error from the
compilers to such files.
Also, when --no-complete-output is not used, the file .stdout is output
before the file .stderr.

Change-Id: I1e54005645f3d16c34e937cc8935bed72e9fc7dd