Handle EX command without a Dep_File or Obj_File argument.
In the case of GPRremote and more generally when compiling without a
language there is no object file or dependency file specified. We do
not want to have an exception here, and so in this case we just do
not try to send back the corresponding files to the build master after
the end of the command execution.

For OB10-014.

Change-Id: I0765e6f30769bf4bda639b6ed76406d72d956ccd