Get_Context, make some parameter optional.
The Get_Context has been fixed to stay compatible with old version
of GPRbuild. This call is used by GPRslave to get the context from
the build master. It must be resilient as much as possible to change
of protocol. With this fix GPRbuild will receive a proper message
saying that the slave is not compatible instead of having GPRslave
raising an exception.

Note that this is only to get a proper message, in no way a slave
will accept to work for a non compatible build master.

Fixes Q217-028.

Change-Id: I0c5ac4b7fd7eabb243ee4768cb5ff7cfcf46ce1d