Q324-020 fix lister & archiver names on windows multiarch
When both 32 and 64 bit toolchains are installed on windows and
32bit gprconfig is called on windows with --target=x86_64-windows
we end up with wrong chunks in the cgpr file :

   for Object_Lister use ("x86_64-pc-mingw32-nm", "-g", "--no-demangle");
   ...
   for Archive_Builder  use ("x86_64-pc-mingw32-ar", "cr");
   ...
   for Archive_Indexer  use ("x86_64-pc-mingw32-ranlib");

These executables do not exist with such setup.

Similar thing happens in reverse situation - 64bit gprconfig called
for --target=x86-windows :

   for Object_Lister use ("i686-pc-mingw32-nm", "-g", "--no-demangle");
   ...
   for Archive_Builder  use ("i686-pc-mingw32-ar", "cr");
   ...
   for Archive_Indexer  use ("i686-pc-mingw32-ranlib");

This creates problems when libraries are involved, since corresponding
executables cannot be found.

Update linker.xml to avoid such problems, add pairs of 32 & 64 bit
windows to <host> tags of corresponding configuration chunks instead
of just one of the two.

Change-Id: Iefd5dfe820d242af4834a2f847da263fedd3ad2d