Fix a small race condition.
We want to send the AK (acknowledge for the compilation job) before
registering the job itself. Not doing that there is a very small
chance that the job is run and output is sent before the ACK is
received.

This makes the counter negative which raises an exception.

For Q328-036.

Change-Id: Ib4d9bdbcdc0c30777716532f01078dc8f6e634a4