Move ar/ranlib handling to new scheme on vx6
Appart from the pure reorg, this also changes arm-vx6 and e500v2-vx6 to
use the same settings as ppc and x86: before 6.4, WRS ar/ranlib; after
6.4, ours. For arm-vx6, this changes nothing as the platform was not
supported on 6.4. For e500v2-vx6, the situation is no different from
the one on ppc, so this change fixes a potential issue.

Change-Id: Iac10b836aeb26b4d08aaf33416e6d1a1f3d55699
TN: O525-029