More fixes for gprls -V -P (Q412-018)
When there is an ALI file in an extended project, but no object file,
in gprls, take the existing ALI file, not the one that should be
created in the ultimate extending project.

A library instantiation is a spec file, but its unit name is deemed
a body; the Sfile should be the spec file.

Change-Id: I532d6b7b2f4df3e647a773e71e782fe2288ed765