Define Ada as a default language.
Define Ada as a default language GPR should fall back to if no languages
are specified in the project. This was implicitly already the case if
gprconfig could find the Ada toolchain, but resulted in no languages defined
for the project if it couldn't. Now Ada is default language regardless of
whether the toolchain was found.

Change-Id: Ide782d614813779a2c30e14935b962dd959a030b
TN: Q523-029