Fix distributed compilation on Windows.
When the build master and the slave does not have the compiler installed
in the same location the build was failing. This is fixed by using the
proper normalization of the compiler path on Windows to ensure that
the paths for runtime files are properly translated in the mapping file.

For Q621-033.

Change-Id: Id5beaffd4ce60ff2aa36732cb63120f8b27a5c18