Concatenation of list attributes (Q829-035)
A new attribute property "configuration concatenable" is introduced for
list attributes.

For the list attributes that have this property, if their values are not
empty in both the user project and the configuration project, the final
value is the concatenation of the configuration project value with the
user project value.
List attributes whose names end with Switches or Options have been marked
with this property in gpr-attr.adb,

For the other list attributes that do not have this property, if a non empty
value is declared in the user project, this is the final value, even if the
value in the configuration project is not empty.

Documentation will be added later.

Change-Id: Ie7f330b108859eb7d4f78996f296d914f4511870