vx6: pass -L options on non-Ada projects only
These options would be useless for programs that have Ada in them,
since these are passed by specs file passed to gcc by a pragma
Linker_Options in system; and would be harmless if it were for
x86-vx6: on this target, to discriminate between vxsim and real
target, we only have an option on command line, and gprconfig/gprbuild
do not have a way to look for such an option on the command line. All
in all, we do not pass them if there is Ada in the project.

Change-Id: I3f70a6586bfc7484c14c389aa597b7995f6bc2c4
TN: QB01-026