Excluded file that should not be (QC02-006)
When several project share the same source directory and
1) one of them excludes an Ada source
and
2) another one indicates a non standard naming for this same Ada source
the compiler may be unable to compile, because the mapping indicates
that the source is excluded, but it should not be so.

Example:

Ada sources:
   main.adb
   pkg_v3.ads
   pkg.adb

main.gpr:
with "prj.gpr";
project Main is
   for Excluded_Source_Files use ("pkg_v3.ads", "pkg.adb");
   for Main use ("main.adb");
end Main;

prj.gpr:
project Prj is
   for Excluded_Source_Files use ("main.adb");

   package Naming is
      for Spec ("pkg") use "pkg_v3.ads";
   end Naming;
end Prj;

Building project Main should not result in a compilation error.

Change-Id: I6cb817d86a7bcd57a064bab10c37e6c116b2cc1d