Specify included dirs for C and C++.
A new configuration attribute Include_Switches_Via_Spec is added
and used to specify how included directories are specified to the
C or C++ compiler, so that the length of the command line is not
two long for the underlying compilers (cc1 and cc1plus).

Some code clean-up and refactoring.

For QC20-038.

Change-Id: I5c2ff1e1d212b37fcfc21c5be9307f5fd073ea00