Specify included dirs for C and C++.
A new configuration attribute Include_Switches_Via_Spec is added
and used to specify how included directories are specified to the
C or C++ compiler, so that the length of the command line is not
two long for the underlying compilers (cc1 and cc1plus).

Some code clean-up and refactoring.

For QC20-038.

Change-Id: I82e92bf66e964aacc7fb42f216c0475b543ef2d2