Refactor rpath support in gprbuild and gprlib
The actual rpath calculation to use for a given directory path is
now put in common in GPR.Util.

The Rpath_Origin is thus now passed in the exchange file as it is
now needed by gprlib.

This fixes invalid rpath used in shared libraries on Windows when
cross-compiling, and ensures a much more homogeneous behavior between
gprbuild and gprlib.

Change-Id: I44fe0d1d984307c0d5bd893d5313191972b4d1a5
TN: R216-036