R409-047 fix closure computation of short subunit names
Add range check in comparison of unit name and name of
parent unit of a subunit. Get_Closures was crashing if
ALI file for a unit with a really long name had among
D lines a subunit of another unit, whose name length
was shorter than the first unit.

Also add a debug output with exception info for
cases when closure cannot be computed due to a crash.

Change-Id: I359e18cf90fa622f87d13e584480ebc88cdbd274