R424-008 fix crash on reference to missing project
When Allow_Invalid_External flag is set to Silent or Warning
and there is a problem with an external declaration do not try
to take the first value of the type as a value for variable if
there are missing withed projects. The type in question may be
from a missing project and this will result in a crash.

Change-Id: Ice33e3aca146c8835196440bf40d74bee393a2b9