Only check for predefined parameter names in --config
Do not treat anything before the colon as a parameter name,
this leads to a wrong behaviour on windows when a full runtime
name is specified and the drive letter is parsed as a parameter.
Instead check only for language, version, runtime, path and name,
and treat anything else as a parameter without a name.

for R530-007 Q309-038

Change-Id: I67dfa159ce845f0dd18d714b0e07ccdfcb010544