R530-028 Handle new artifacts in gprclean + cleanup tmp files logic.
Artifacts from 1st and 2nd stages of static SALs linking are now removed
by gprclean.

Also, make more use of the temporary files mechanism and fix it:
the code to delete temp files was checking on Opt.Keep_Temp_Files
which was not set in the command line as expected.
Debug flag "dn" is now merely another way to set this option.

Change-Id: I4256012afac3050174bba64ada33fe1aef13499e