R323-019 Get rid of the 2 sec. tolerance on time stamp comparisons.
When a file system uses a granularity G for time stamps, it just
means that if 2 files are created at T1 and T2 and receive
time stamps TS1 and TS2, then we may have TS1 = TS2 if |T1-T2| <= G.
But if TS1 /= TS2 then it is always true that T1 /= T2.
Adding a tolerance G to the TS comparison doesn't make sense.

Change-Id: I04c09d39a888ed74fc57bf021f4079e4d1d70cc8