vx7r2cert: link with -nostdlib
In cert configurations, we need to let the user control precisely
which libraries are to be linked in the executable; only libgnat and
libgnarl would be added by default with --RTS. So, with cert rts,
always pass -nostdlib. This in particular makes sure that libgcc,
libgcc_eh and crt* files are not used by default.

Change-Id: Icd8c5891319769c39e85035ca9d9a1f4c6c74261
TN: T813-071