Fix partial link when gcc is built with --enable-default-pie
When gcc is built with --enable-default-pie, it automatically passes
-pie to linker commands, unless gcc itself is run with -no-pie. However,
building auto-initialized static libraries, gprbuild passes -Wl,-r to
gcc, which means that in such configurations, gcc will run the linker
with both -pie and -r.

These options are incompatible. As a consequence, building static
auto-initialized libraries with gprbuild is broken when gcc was built
with --enable-default-pie, i.e. with the system compiler on most recent
GNU/Linux distributions.

This commit fixes this issue adding -no-pie to the partial link command,
except for GNAT releases prior to 17.1, which do not support -no-pie.

Change-Id: Icd5a6405ac430957ac618a5ea4e01fca94462d80
TN: S416-025