Forcing compilation with -mcmodel=large for x86_64-vx7r2 platforms.
This modification is a workaround because vxworks loader does not manage
modules made with mixed memory model on x86_64 platforms..
Since the large memory model has been used for the ada runtimes,
The application should be using the same model.
This will be reverted if we decide to revert the runtime to another
memory model.

TN: U407-053
tested-with: $ anod test acats-4 --target=x86_64-vx7r2,7.2,qemu \
   --qualifier=rts=kernel
 $ anod test aunit --target=x86_64-vx7r2,7.2,qemu --qualifier=rts=kernel
 $ anod test aunit --target=ppc64-vx7r2,7.2,qemu --qualifier=rts=kernel
Change-Id: Iaed5afc23a54a55792d32df7e1fb2a2315abb6dd